requirements.txt
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
autopep8==1.6.0
backports.zoneinfo==0.2.1
cached-property==1.5.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
coverage==6.3.2
cryptography==36.0.2
defusedxml==0.7.1
Deprecated==1.2.13
Django==3.2
django-admin-interface==0.19.0
django-allauth==0.50.0
django-colorfield==0.6.3
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-redis==5.2.0
django-rest-auth==0.9.5
django-silk==4.3.0
django-url-filter==0.3.15
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
drf-spectacular==0.22.0
drf-yasg==1.20.0
enum-compat==0.0.3
gprof2dot==2021.2.21
gunicorn==20.1.0
idna==3.3
importlib-resources==5.7.1
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.1
jsonschema==4.4.0
MarkupSafe==2.1.1
oauthlib==3.2.0
openapi==1.1.0
packaging==21.3
Pillow==9.1.0
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.8
pyrsistent==0.18.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.1
PyYAML==6.0
redis==4.2.2
requests==2.27.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
wrapt==1.14.0
zipp==3.8.0
